requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
einops>=0.6.1
scikit-learn>=1.2.0
pandas>=2.0.0
